P5/requirements.txt
numpy
scipy
scikit-learn
pandas
liac-arff
